Fix bug in zip accessor, Change writer to outputstream
The jackson library by default closes a stream after the mapper call to writevalue; this resulted in the ZipOutputStream being closed after writing the manifest and throwing an exception when copying any files remaining afterwards.

Writer replaced with OutputStream to allow for eventual addition of writing binary files, not just text (adding media to the RC).